+++ lib/slack/web/api/helpers/files.rb
@@ -32,6 +32,10 @@ module Slack
               raise ArgumentError, "Required argument :#{param} missing" if params[param].nil?
             end
 
+            channels = Array(params[:channels]).map do |channel|
+              conversations_id(channel: channel)['channel']['id']
+            end.uniq.join(',')
+
             content = params[:content]
             title = params[:title] || params[:filename]
 
@@ -55,7 +59,8 @@ module Slack
             end
 
             # Complete the upload.
-            complete_upload_request_params = params.slice(:channels, :initial_comment, :thread_ts)
+            complete_upload_request_params = params.slice(:initial_comment, :thread_ts)
+            complete_upload_request_params[:channels] = channels
             complete_upload_request_params[:files] = [{ id: file_id, title: title }].to_json
 
             files_completeUploadExternal(complete_upload_request_params)