+++ spec/slack/web/api/endpoints/custom_specs/files_spec.rb
@@ -36,6 +36,27 @@ RSpec.describe Slack::Web::Api::Endpoints::Files do
     end
   end
 
+  context 'when a channel is referenced by name', vcr: { cassette_name: 'web/files_upload_v2_with_channel_name' } do
+    before do
+      allow(client).to receive(:conversations_list).and_yield(
+        Slack::Messages::Message.new(
+          'channels' => [{
+            'id' => 'C08AZ76CA4V',
+            'name' => 'general'
+          }]
+        )
+      )
+    end
+
+    it 'translates the channel name to an ID and completes the upload' do
+      expect(client.files_upload_v2(
+        filename: 'test.txt',
+        content: 'Test File Contents',
+        channels: '#general'
+      ).files.size).to eq 1
+    end
+  end
+
   context 'when using a list for channels', vcr: { cassette_name: 'web/files_upload_v2_with_channels_list' } do
     it 'completes the upload' do
       expect(client.files_upload_v2(
@@ -56,6 +77,27 @@ RSpec.describe Slack::Web::Api::Endpoints::Files do
         initial_comment: 'initial_comment',
         thread_ts: '1738331914.958599',
         snippet_type: 'text'
+      ).files.size).to eq 1
+    end
+  end
+
+  context 'with an array of channels', vcr: { cassette_name: 'web/files_upload_v2_with_channels_list' } do
+    before do
+      allow(client).to receive(:conversations_list).and_yield(
+        Slack::Messages::Message.new(
+          'channels' => [{
+            'id' => 'C08BHPZBZ8A',
+            'name' => 'general'
+          }]
+        )
+      )
+    end
+
+    it 'completes the upload' do
+      expect(client.files_upload_v2(
+        filename: 'test.txt',
+        content: 'Test File Contents',
+        channels: ['C08AZ76CA4V', '#general']
       ).files.size).to eq 1
     end
   end